requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
distro==1.8.0
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
identify==2.5.33
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
nodeenv==1.8.0
openai==1.1.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
platformdirs==3.11.0
pre-commit==3.6.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.25.0
wcwidth==0.2.9
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
